move sidebar components to /sidebar
