+++ apps/web/src/app/dashboard/layout.tsx
@@ -1,6 +1,6 @@
 import { cookies } from "next/headers";
 
-import { AppSidebar } from "@/app/dashboard/components/app-sidebar";
+import { AppSidebar } from "@/app/dashboard/components/sidebar/app-sidebar";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 
 export default async function Layout({
